LC-409 Do not show counters when zero sources
